pyproject.toml
[project]
name = "e-commerce-ai-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.2",
    "langchain>=1.0.1",
    "langchain-mcp-adapters>=0.1.11",
    "langchain-openai>=1.0.0",
    "langgraph>=1.0.0",
    "mcp>=1.18.0",
    "pydantic>=2.12.3",
    "python-dotenv>=1.1.1",
]
